supervised classification has been added as an option to web service side. Moreover, word vectors has been used as paragraph vector calculation.
-- the avarage of word vectors must be modified.
-- some refactoring is need.
-- umls:similarity should be checked again (perl script)

pairs_berna has been added to the system for testing. --
some comment lines are removed from the classes to prevent so much input in the terminal.

As next step, log file should be added to the system and system should be managed by using this logs.
Some refactoring works are needed for better usage of the system